refactor: split oversized files under 500-line limit
Split Zod schemas from mcp/types.ts (508→458 lines) into mcp/schemas.ts.
Split factory functions from d1/client.ts (544→498 lines) into d1/factory.ts.
Backward-compatible re-exports preserved at original import paths.
