+++ worker/lib/d1/client.ts
@@ -8,6 +8,13 @@ import type {
   D1PreparedStatement,
 } from "@cloudflare/workers-types";
 import { logger } from "../global-logger";
+import { stripSqlComments } from "./factory";
+
+export {
+  createD1Client,
+  createD1ReadClient,
+  createD1WriteClient,
+} from "./factory";
 
 // ============================================================================
 // Error Types
@@ -486,59 +493,6 @@ export class D1Client {
 }
 
 // ============================================================================
-// SQL Helpers
+// SQL helper imported from factory.ts to keep this file under 500 lines.
+// Factory functions (createD1Client, etc.) are also in factory.ts.
 // ============================================================================
-
-/**
- * Strip SQL single-line comment lines (-- ...) and blank lines from a SQL string.
- * Preserves inline comments (e.g. `SELECT 1 -- test` remains `SELECT 1`).
- * Preserves block comments (/* ... *​/).
- */
-function stripSqlComments(sql: string): string {
-  return sql
-    .split("\n")
-    .filter((line) => {
-      const trimmed = line.trim();
-      return trimmed.length > 0 && !trimmed.startsWith("--");
-    })
-    .join("\n");
-}
-
-// ============================================================================
-// Factory Functions
-// ============================================================================
-
-/**
- * Create a D1 client with default configuration
- */
-export function createD1Client(
-  db: D1Database,
-  config?: D1ClientConfig,
-): D1Client {
-  return new D1Client(db, config);
-}
-
-/**
- * Create a D1 client optimized for reads (uses sessions for replication)
- */
-export function createD1ReadClient(
-  db: D1Database,
-  bookmark?: string,
-): D1Client {
-  return new D1Client(db, {
-    useSessions: true,
-    sessionBookmark: bookmark || "first-unconstrained",
-    enableRetries: true,
-  });
-}
-
-/**
- * Create a D1 client optimized for writes
- */
-export function createD1WriteClient(db: D1Database): D1Client {
-  return new D1Client(db, {
-    useSessions: true,
-    sessionBookmark: "first-primary", // Forces primary for immediate consistency
-    enableRetries: true,
-  });
-}
